Move SecretsStore interface
